Improve Packager efficiency.
