+++ src/main/java/com/lothrazar/cyclic/block/packager/TilePackager.java
@@ -6,18 +6,13 @@ import com.lothrazar.cyclic.block.battery.TileBattery;
 import com.lothrazar.cyclic.capability.CustomEnergyStorage;
 import com.lothrazar.cyclic.capability.ItemStackHandlerWrapper;
 import com.lothrazar.cyclic.registry.TileRegistry;
-import java.util.HashMap;
-import java.util.Map;
 import net.minecraft.block.BlockState;
 import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.entity.player.PlayerInventory;
 import net.minecraft.inventory.container.Container;
 import net.minecraft.inventory.container.INamedContainerProvider;
-import net.minecraft.item.Item;
 import net.minecraft.item.ItemStack;
 import net.minecraft.item.crafting.ICraftingRecipe;
-import net.minecraft.item.crafting.IRecipeType;
-import net.minecraft.item.crafting.Ingredient;
 import net.minecraft.item.crafting.RecipeManager;
 import net.minecraft.nbt.CompoundNBT;
 import net.minecraft.tileentity.ITickableTileEntity;
@@ -25,7 +20,6 @@ import net.minecraft.util.Direction;
 import net.minecraft.util.text.ITextComponent;
 import net.minecraft.util.text.StringTextComponent;
 import net.minecraftforge.common.ForgeConfigSpec.IntValue;
-import net.minecraftforge.common.Tags;
 import net.minecraftforge.common.capabilities.Capability;
 import net.minecraftforge.common.util.LazyOptional;
 import net.minecraftforge.energy.CapabilityEnergy;
@@ -45,15 +39,18 @@ public class TilePackager extends TileEntityBase implements INamedContainerProvi
   public static final int TICKS = 10;
   CustomEnergyStorage energy = new CustomEnergyStorage(MAX, MAX);
   private final LazyOptional<IEnergyStorage> energyCap = LazyOptional.of(() -> energy);
-  ItemStackHandler inputSlots = new ItemStackHandler(1);
+  ItemStackHandler inputSlots = new ItemStackHandler(1) {
+
+    @Override
+    public boolean isItemValid(int slot, ItemStack stack) {
+      return isItemStackValid(stack);
+    }
+  };
   ItemStackHandler outputSlots = new ItemStackHandler(1);
   private final ItemStackHandlerWrapper inventory = new ItemStackHandlerWrapper(inputSlots, outputSlots);
   private final LazyOptional<IItemHandler> inventoryCap = LazyOptional.of(() -> inventory);
   private int burnTimeMax = 0; //only non zero if processing
   private int burnTime = 0; //how much of current fuel is left
-  private static final Map<Item, ICraftingRecipe> fourItemRecipeCache = new HashMap<>();
-  private static final Map<Item, ICraftingRecipe> nineItemRecipeCache = new HashMap<>();
-  private static final Map<ICraftingRecipe, Integer> ingredientsInRecipeCache = new HashMap<>();
 
   public TilePackager() {
     super(TileRegistry.PACKAGER.get());
@@ -88,12 +85,12 @@ public class TilePackager extends TileEntityBase implements INamedContainerProvi
     if (world == null) {
       return;
     }
-    final ICraftingRecipe recipe = getRecipeForItemStack(world.getRecipeManager(), stack);
+    final ICraftingRecipe recipe = UtilPackager.getRecipeForItemStack(world.getRecipeManager(), stack);
     if (recipe == null) {
       return;
     }
     if (outputSlots.insertItem(0, recipe.getRecipeOutput().copy(), true).isEmpty()) {
-      final int total = ingredientsInRecipeCache.get(recipe);
+      final int total = UtilPackager.getIngredientsInRecipe(recipe);
       final ItemStack output = recipe.getRecipeOutput().copy();
       ModCyclic.LOGGER.info("Packager recipe match of size " + total + " producing -> " + output);
       //consume items, produce output
@@ -103,107 +100,12 @@ public class TilePackager extends TileEntityBase implements INamedContainerProvi
     }
   }
 
-  public static ICraftingRecipe getRecipeForItemStack(final RecipeManager recipeManager, final ItemStack itemStack) {
-    if (itemStack.getCount() >= 9) {
-      if (nineItemRecipeCache.isEmpty()) {
-        buildRecipeCaches(recipeManager);
-      }
-      final ICraftingRecipe recipe = nineItemRecipeCache.get(itemStack.getItem());
-      if (recipe != null) {
-        return recipe;
-      }
-    }
-    if (itemStack.getCount() >= 4) {
-      if (fourItemRecipeCache.isEmpty()) {
-        buildRecipeCaches(recipeManager);
-      }
-      return fourItemRecipeCache.get(itemStack.getItem());
-    }
-    return null;
-  }
-
-  public static boolean isRecipeValid(final ICraftingRecipe recipe) {
-    final ItemStack recipeOutput = recipe.getRecipeOutput();
-    if (recipeOutput.getMaxStackSize() == 1 || recipeOutput.getCount() != 1) {
+  public boolean isItemStackValid(final ItemStack itemStack) {
+    if (world == null) {
       return false;
     }
-    Ingredient mainIngredient = null;
-    ItemStack mainIngredientStack = null;
-    int count = 0;
-    for (final Ingredient ingredient : recipe.getIngredients()) {
-      if (ingredient == Ingredient.EMPTY) {
-        continue;
-      }
-      final ItemStack[] matchingStacks = ingredient.getMatchingStacks();
-      if (matchingStacks.length == 0) {
-        continue;
-      }
-      else if (matchingStacks.length > 1) {
-        return false;
-      }
-      final ItemStack matchingStack = matchingStacks[0];
-      if (mainIngredient != null && !mainIngredient.test(matchingStack)) {
-        return false;
-      }
-      mainIngredient = ingredient;
-      mainIngredientStack = matchingStack;
-      count++;
-    }
-    if (mainIngredient == null) {
-      return false;
-    }
-    boolean outIsStorage = recipeOutput.getItem().isIn(Tags.Items.STORAGE_BLOCKS);
-    final Item mainIngredientItem = mainIngredientStack.getItem();
-    boolean inIsIngot = mainIngredientItem.isIn(Tags.Items.INGOTS);
-    if (!outIsStorage && inIsIngot) {
-      return false;
-    }
-    return true;
-  }
-
-  public static void buildRecipeCaches(final RecipeManager recipeManager) {
-    recipeLoop: for (final ICraftingRecipe recipe : recipeManager.getRecipesForType(IRecipeType.CRAFTING)) {
-      final ItemStack recipeOutput = recipe.getRecipeOutput();
-      if (recipeOutput.getMaxStackSize() == 1 || recipeOutput.getCount() != 1) {
-        continue;
-      }
-      Ingredient mainIngredient = null;
-      ItemStack mainIngredientStack = null;
-      int count = 0;
-      for (final Ingredient ingredient : recipe.getIngredients()) {
-        if (ingredient == Ingredient.EMPTY) {
-          continue;
-        }
-        final ItemStack[] matchingStacks = ingredient.getMatchingStacks();
-        if (matchingStacks.length != 1) {
-          continue recipeLoop;
-        }
-        final ItemStack matchingStack = matchingStacks[0];
-        if (mainIngredient != null && !mainIngredient.test(matchingStack)) {
-          continue recipeLoop;
-        }
-        mainIngredient = ingredient;
-        mainIngredientStack = matchingStack;
-        count++;
-      }
-      if (mainIngredient == null) {
-        continue;
-      }
-      boolean outIsStorage = recipeOutput.getItem().isIn(Tags.Items.STORAGE_BLOCKS);
-      final Item mainIngredientItem = mainIngredientStack.getItem();
-      boolean inIsIngot = mainIngredientItem.isIn(Tags.Items.INGOTS);
-      if (!outIsStorage && inIsIngot) {
-        continue;
-      }
-      if (count == 4) {
-        fourItemRecipeCache.put(mainIngredientItem, recipe);
-        ingredientsInRecipeCache.put(recipe, 4);
-      }
-      else if (count == 9) {
-        nineItemRecipeCache.put(mainIngredientItem, recipe);
-        ingredientsInRecipeCache.put(recipe, 9);
-      }
-    }
+    final RecipeManager recipeManager = world.getRecipeManager();
+    return UtilPackager.isItemStackValid(recipeManager, itemStack);
   }
 
   @Override
